+++ packages/formats/src/index.ts
@@ -1,1 +1,3 @@
 export * from "./rgbeffects";
+export * from "./fseq";
+export * from "./xsq";